# api/openapi.yaml
openapi: 3.0.3
info:
  title: QRATUM Platform API
  description: |
    QRATUM (Quantum Resource Allocation, Tensor Analysis, and Unified Modeling) Platform API.
    
    This API provides endpoints for:
    - Job submission (quantum simulations, tensor analysis)
    - Real-time status monitoring (WebSocket support)
    - Results retrieval and visualization
    - Resource allocation dashboards
    
    ## Security
    All endpoints require OAuth2/OIDC authentication. API keys are supported for
    service-to-service communication.
    
    ## Compliance
    - DO-178C Level A certification maintained
    - NIST 800-53 security controls enforced
    - CMMC 2.0 Level 2 compliant
    - DFARS requirements met
  version: 1.0.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: QRATUM Team
    url: https://github.com/robertringler/QRATUM
servers:
  - url: https://api.qratum.io/v1
    description: Production server
  - url: https://staging-api.qratum.io/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Local development
...
paths:
...
  /resources:
    get:
      tags:
        - Resources
      summary: Get resource allocation dashboard
      description: |
        Get current resource allocation and utilization metrics.
        Includes cluster status, GPU availability, and queue depth.
      operationId: getResources
      security:
        - bearerAuth: []
        - apiKey: []
      responses:
        '200':
          description: Resource allocation data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceDashboard'
  /resources/clusters:
    get:
      tags:
        - Resources
      summary: List available clusters
      description: List all available compute clusters (EKS, GKE, AKS)
      operationId: listClusters
      security:
        - bearerAuth: []
        - apiKey: []
      responses:
        '200':
          description: List of clusters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClusterListResponse'
  /resources/quotas:
    get:
      tags:
        - Resources
      summary: Get resource quotas
      description: Get current user's resource quotas and usage
      operationId: getQuotas
      security:
        - bearerAuth: []
        - apiKey: []
      responses:
        '200':
          description: Resource quotas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuotaResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from `/auth/token` endpoint.
        Include as `Authorization: Bearer <token>` header.
    
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key for service-to-service communication.
        Generated from QRATUM dashboard or via Vault.
    
    oauth2:
      type: oauth2
      description: OAuth2 with OIDC for user authentication
      flows:
        authorizationCode:
          authorizationUrl: https://auth.qratum.io/authorize
          tokenUrl: https://auth.qratum.io/token
          scopes:
            read:jobs: Read job information
            write:jobs: Submit and manage jobs
            read:resources: Read resource information
            admin: Administrative access
...
  schemas:
...
    ResourceDashboard:
      type: object
      required:
        - timestamp
        - clusters
        - queue_depth
        - utilization
      properties:
        timestamp:
          type: string
          format: date-time
        clusters:
          type: array
          items:
            $ref: '#/components/schemas/ClusterSummary'
        queue_depth:
          type: integer
        jobs_running:
          type: integer
        jobs_queued:
          type: integer
        utilization:
          type: object
          properties:
            cpu:
              type: number
            memory:
              type: number
            gpu:
              type: number
    ClusterSummary:
      type: object
      required:
        - cluster_id
        - provider
        - region
        - status
      properties:
        cluster_id:
          type: string
        name:
          type: string
        provider:
          type: string
          enum: [aws, gcp, azure]
        region:
          type: string
        status:
          type: string
          enum: [online, degraded, offline]
        nodes:
          type: integer
        gpus_available:
          type: integer
        gpus_total:
          type: integer
    ClusterListResponse:
      type: object
      required:
        - clusters
      properties:
        clusters:
          type: array
          items:
            $ref: '#/components/schemas/ClusterSummary'
    QuotaResponse:
      type: object
      required:
        - quotas
        - usage
      properties:
        quotas:
          type: object
          properties:
            max_concurrent_jobs:
              type: integer
            max_gpu_hours_per_month:
              type: integer
            max_storage_gb:
              type: integer
        usage:
          type: object
          properties:
            concurrent_jobs:
              type: integer
            gpu_hours_this_month:
              type: number
            storage_gb:
              type: number